chore: split nilauth client
