+++ src/nilauth/client.ts
@@ -2,83 +2,31 @@ import { randomBytes } from "node:crypto";
 import { secp256k1 } from "@noble/curves/secp256k1";
 import { sha256 } from "@noble/hashes/sha256";
 import { bytesToHex } from "@noble/hashes/utils";
-import { Effect as E, Schedule, pipe } from "effect";
+import { Effect as E, pipe } from "effect";
 import { Temporal } from "temporal-polyfill";
-import z from "zod";
 import { NucTokenBuilder } from "#/builder";
-import { type NucTokenEnvelope, NucTokenEnvelopeSchema } from "#/envelope";
+import type { NucTokenEnvelope } from "#/envelope";
 import type { Keypair } from "#/keypair";
 import { log } from "#/logger";
+import { type NilauthError, sendRequest } from "#/nilauth/request-sender";
+import {
+  type CreateTokenResponse,
+  CreateTokenResponseSchema,
+  type LookupRevokedTokenResponse,
+  LookupRevokedTokenResponseSchema,
+  type NilauthAboutResponse,
+  NilauthAboutResponseSchema,
+  type NilauthHealthResponse,
+  NilauthHealthResponseSchema,
+  type SubscriptionCostResponse,
+  SubscriptionCostResponseSchema,
+  type ValidatePaymentRequest,
+} from "#/nilauth/types";
 import type { Payer } from "#/payer/client";
-import type { TxHash } from "#/payer/types";
 import { Did, InvocationBody, REVOKE_COMMAND } from "#/token";
 
 const PAYMENT_TX_RETRIES = [1000, 2000, 3000, 5000, 10000, 10000, 10000];
 const TX_RETRY_ERROR_CODE = "TRANSACTION_NOT_COMMITTED";
-
-export const NilauthHealthResponseSchema = z.literal("OK");
-export type NilauthHealthResponse = z.infer<typeof NilauthHealthResponseSchema>;
-
-export const BuildSchema = z
-  .object({
-    commit: z.string(),
-    timestamp: z.string(),
-  })
-  .transform(({ commit, timestamp }) => ({
-    commit,
-    timestamp: Temporal.Instant.from(timestamp),
-  }));
-
-export const NilauthAboutResponseSchema = z
-  .object({
-    started: z.string(),
-    public_key: z.string(),
-    build: BuildSchema,
-  })
-  .transform(({ started, public_key, build }) => ({
-    started: Temporal.Instant.from(started),
-    publicKey: public_key,
-    build,
-  }));
-export type NilauthAboutResponse = z.infer<typeof NilauthAboutResponseSchema>;
-
-export const SubscriptionCostResponseSchema = z
-  .object({
-    cost_unils: z.number(),
-  })
-  .transform(({ cost_unils }) => cost_unils);
-export type SubscriptionCostResponse = z.infer<
-  typeof SubscriptionCostResponseSchema
->;
-
-export const CreateTokenResponseSchema = z.object({
-  token: NucTokenEnvelopeSchema,
-});
-export type CreateTokenResponse = z.infer<typeof CreateTokenResponseSchema>;
-
-export const RevokedTokenSchema = z
-  .object({
-    token_hash: z.string(),
-    revoked_at: z.string(),
-  })
-  .transform(({ token_hash, revoked_at }) => ({
-    tokenHash: token_hash,
-    revokedAt: revoked_at,
-  }));
-export type RevokedToken = z.infer<typeof RevokedTokenSchema>;
-
-export const LookupRevokedTokenResponseSchema = z.object({
-  revoked: z.array(RevokedTokenSchema),
-});
-export type LookupRevokedTokenResponse = z.infer<
-  typeof LookupRevokedTokenResponseSchema
->;
-
-type ValidatePaymentRequest = {
-  tx_hash: TxHash;
-  payload: string;
-  public_key: string;
-};
 
 export class NilauthClient {
   constructor(
@@ -185,7 +133,7 @@ export class NilauthClient {
       url: `${this.baseUrl}/api/v1/payments/validate`,
       timeout: this.timeout,
       retryDelays: PAYMENT_TX_RETRIES,
-      retryWhile: (error) => error.code === TX_RETRY_ERROR_CODE,
+      retryWhile: (error: NilauthError) => error.code === TX_RETRY_ERROR_CODE,
       init: {
         method: "POST",
         headers: { "Content-Type": "application/json" },
@@ -305,108 +253,4 @@ export class NilauthClient {
       E.runPromise,
     );
   }
-}
-
-type NilauthRequest = {
-  url: string;
-  timeout: number;
-  init?: RequestInit;
-  retryDelays?: number[];
-  retryWhile?: (error: NilauthError) => boolean;
-};
-
-const NilauthResponseSchema = z.union([
-  z.string(),
-  z.record(z.unknown()),
-  z.null(),
-]);
-type NilauthResponse = z.infer<typeof NilauthResponseSchema>;
-
-class NilauthError extends Error {
-  constructor(
-    public readonly code: string,
-    message: string,
-  ) {
-    super(`${code}: ${message}`);
-  }
-}
-const NilauthErrorSchema = z
-  .object({
-    error_code: z.string(),
-    message: z.string(),
-  })
-  .transform(
-    ({ error_code, message }) => new NilauthError(error_code, message),
-  );
-
-async function sendRequest(request: NilauthRequest): Promise<unknown> {
-  const {
-    url,
-    timeout,
-    init,
-    retryDelays = [],
-    retryWhile = (_) => false,
-  } = request;
-  const maxRetries = retryDelays.length;
-  return pipe(
-    E.retry(
-      pipe(
-        E.tryPromise(() => fetchWithTimeout(url, timeout, init)),
-        E.andThen((response) => parseNilauthResponse(response)),
-      ),
-      pipe(
-        Schedule.recurs(maxRetries),
-        Schedule.delayed(() => retryDelays.shift() ?? 0),
-        Schedule.whileInput((error) => {
-          if (error instanceof NilauthError && retryWhile(error)) {
-            log(`retrying: ${error}`);
-            return true;
-          }
-          if (error instanceof Error && error.cause === "timeout") {
-            log(`retrying: ${error.cause}`);
-            return true;
-          }
-          return false;
-        }),
-      ),
-    ),
-    E.runPromise,
-  );
-}
-
-async function fetchWithTimeout(
-  url: string,
-  timeout: number,
-  init?: RequestInit,
-): Promise<Response> {
-  const timeoutPromise = new Promise((_, reject) =>
-    setTimeout(() => reject("timeout"), timeout),
-  );
-  return (await Promise.race([fetch(url, init), timeoutPromise])) as Response;
-}
-
-function parseNilauthResponse(
-  response: Response,
-): E.Effect<NilauthResponse, Error> {
-  const contentType = response.headers.get("content-type");
-  if (!contentType) return E.fail(new Error("content-type not found"));
-  if (contentType.includes("text/plain")) {
-    return pipe(
-      E.tryPromise(() => response.text()),
-      E.flatMap((body) => E.try(() => NilauthResponseSchema.parse(body))),
-    );
-  }
-  if (contentType === "application/json") {
-    if (!response.ok)
-      return pipe(
-        E.tryPromise(() => response.json()),
-        E.flatMap((body) => E.try(() => NilauthErrorSchema.parse(body))),
-        E.flatMap((body) => E.fail(body)),
-      );
-    return pipe(
-      E.tryPromise(() => response.json()),
-      E.flatMap((body) => E.try(() => NilauthResponseSchema.parse(body))),
-    );
-  }
-  return E.fail(new Error("unsupported content-type"));
 }
